Add files via upload
This was my first test run with developing a long-short term neural network in python. It uses a single time series as input data to predict a single time series of output data. The data is based around monthly milk production. Comments on how it works can be seen in the code file itself. Developed for Capstone Design ECSE-4900 at Rensselaer Polytechnic Institute.